+++ tidy3d/__init__.py
@@ -152,7 +152,13 @@ from .components.heat_charge.source import HeatFromElectricSource, HeatSource, U
 from .components.heat_charge_spec import ConductorSpec, FluidSpec, InsulatorSpec, SolidSpec
 
 # lumped elements
-from .components.lumped_element import CoaxialLumpedResistor, LumpedResistor
+from .components.lumped_element import (
+    AdmittanceNetwork,
+    CoaxialLumpedResistor,
+    LinearLumpedElement,
+    LumpedResistor,
+    RLCNetwork,
+)
 
 # medium
 # for docs
@@ -458,8 +464,11 @@ __all__ = [
     "config",
     "__version__",
     "Updater",
-    "LumpedResistor",
+    "AdmittanceNetwork",
     "CoaxialLumpedResistor",
+    "LinearLumpedElement",
+    "LumpedResistor",
+    "RLCNetwork",
     "Scene",
     "StructureStructureInterface",
     "StructureBoundary",